Trait mycelium_bitfield::FromBits

pub trait FromBits<B>: Sized {
    type Error: Display;

    const BITS: u32;

    // Required methods
    fn try_from_bits(bits: B) -> Result<Self, Self::Error>;
    fn into_bits(self) -> B;
}

Trait implemented by values which can be converted to and from raw bits.

This trait is implemented by default for all signed and unsigned integer types, as well as for bools. It can be implemented manually for any user-defined type which has a well-defined bit-pattern representation. For enum types with unsigned integer reprs, it may also be implemented automatically using the enum_from_bits! macro.

Required Associated Types

type Error: Display

The error type returned by Self::try_from_bits when an invalid bit pattern is encountered.

If all bit patterns possible in Self::BITS bits are valid bit patterns for a Self-typed value, this should generally be core::convert::Infallible.

Required Associated Constants

const BITS: u32

The number of bits required to represent a value of this type.

Required Methods

fn try_from_bits(bits: B) -> Result<Self, Self::Error>

Attempt to convert bits into a value of this type.

Returns

Ok(Self) if bits contained a valid bit pattern for a value of this type.

Err(Self::Error) if bits is an invalid bit pattern for a value of this type.

fn into_bits(self) -> B

Convert self into a raw bit representation.

In general, this will be a low-cost conversion (e.g., for enums, this is generally an as cast).

Object Safety

This trait is not object safe.